move modules icons
