--- redisinsight/ui/src/components/base/icons/modules/RediStackDarkLogoIcon.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-import { Icon, IconProps } from 'uiSrc/components/base/icons'
-import RediStackDarkLogoSvg from 'uiSrc/assets/img/modules/redistack/RedisStackLogoDark.svg?react'
-
-export const RediStackDarkLogoIcon = (props: IconProps) => (
-  <Icon icon={RediStackDarkLogoSvg} {...props} isSvg />
-)
