--- redisinsight/ui/src/components/base/icons/modules/RediStackDarkMinIcon.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-import { Icon, IconProps } from 'uiSrc/components/base/icons'
-import RediStackDarkMinSvg from 'uiSrc/assets/img/modules/redistack/RediStackDark-min.svg?react'
-
-export const RediStackDarkMinIcon = (props: IconProps) => (
-  <Icon icon={RediStackDarkMinSvg} {...props} isSvg />
-)
