--- redisinsight/ui/src/components/base/icons/modules/RediStackLightLogoIcon.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-import { Icon, IconProps } from 'uiSrc/components/base/icons'
-import RediStackLightLogoSvg from 'uiSrc/assets/img/modules/redistack/RedisStackLogoLight.svg?react'
-
-export const RediStackLightLogoIcon = (props: IconProps) => (
-  <Icon icon={RediStackLightLogoSvg} {...props} isSvg />
-)
