--- redisinsight/ui/src/components/base/icons/modules/RediStackLightMinIcon.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-import RediStackLightMinLight from 'uiSrc/assets/img/modules/redistack/RediStackLight-min.svg?react'
-import { Icon, IconProps } from 'uiSrc/components/base/icons'
-
-export const RediStackLightMinIcon = (props: IconProps) => (
-  <Icon icon={RediStackLightMinLight} {...props} isSvg />
-)
