--- redisinsight/ui/src/components/base/icons/modules/RedisAIDarkIcon.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-import RedisAIDark from 'uiSrc/assets/img/modules/RedisAIDark.svg?react'
-import { Icon, IconProps } from 'uiSrc/components/base/icons'
-
-export const RedisAIDarkIcon = (props: IconProps) => (
-  <Icon icon={RedisAIDark} {...props} isSvg />
-)
