--- redisinsight/ui/src/components/base/icons/modules/RedisAILightIcon.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-import RedisAILight from 'uiSrc/assets/img/modules/RedisAILight.svg?react'
-import { Icon, IconProps } from 'uiSrc/components/base/icons'
-
-export const RedisAILightIcon = (props: IconProps) => (
-  <Icon icon={RedisAILight} {...props} isSvg />
-)
