--- redisinsight/ui/src/components/base/icons/modules/RedisBloomDarkIcon.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-import RedisBloomDark from 'uiSrc/assets/img/modules/RedisBloomDark.svg?react'
-import { Icon, IconProps } from 'uiSrc/components/base/icons'
-
-export const RedisBloomDarkIcon = (props: IconProps) => (
-  <Icon icon={RedisBloomDark} {...props} isSvg />
-)
