--- redisinsight/ui/src/components/base/icons/modules/RedisBloomLightIcon.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-import RedisBloomLight from 'uiSrc/assets/img/modules/RedisBloomLight.svg?react'
-import { Icon, IconProps } from 'uiSrc/components/base/icons'
-
-export const RedisBloomLightIcon = (props: IconProps) => (
-  <Icon icon={RedisBloomLight} {...props} isSvg />
-)
