--- redisinsight/ui/src/components/base/icons/modules/RedisGears2DarkIcon.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-import RedisGears2Dark from 'uiSrc/assets/img/modules/RedisGears2Dark.svg?react'
-import { Icon, IconProps } from 'uiSrc/components/base/icons'
-
-export const RedisGears2DarkIcon = (props: IconProps) => (
-  <Icon icon={RedisGears2Dark} {...props} isSvg />
-)
