--- redisinsight/ui/src/components/base/icons/modules/RedisGears2LightIcon.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-import RedisGears2Light from 'uiSrc/assets/img/modules/RedisGears2Light.svg?react'
-import { Icon, IconProps } from 'uiSrc/components/base/icons'
-
-export const RedisGears2LightIcon = (props: IconProps) => (
-  <Icon icon={RedisGears2Light} {...props} isSvg />
-)
