--- redisinsight/ui/src/components/base/icons/modules/RedisGearsDarkIcon.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-import RedisGearsDark from 'uiSrc/assets/img/modules/RedisGearsDark.svg?react'
-import { Icon, IconProps } from 'uiSrc/components/base/icons'
-
-export const RedisGearsDarkIcon = (props: IconProps) => (
-  <Icon icon={RedisGearsDark} {...props} isSvg />
-)
