--- redisinsight/ui/src/components/base/icons/modules/RedisGearsLightIcon.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-import RedisGearsLight from 'uiSrc/assets/img/modules/RedisGearsLight.svg?react'
-import { Icon, IconProps } from 'uiSrc/components/base/icons'
-
-export const RedisGearsLightIcon = (props: IconProps) => (
-  <Icon icon={RedisGearsLight} {...props} isSvg />
-)
