--- redisinsight/ui/src/components/base/icons/modules/RedisGraphDarkIcon.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-import RedisGraphDark from 'uiSrc/assets/img/modules/RedisGraphDark.svg?react'
-import { Icon, IconProps } from 'uiSrc/components/base/icons'
-
-export const RedisGraphDarkIcon = (props: IconProps) => (
-  <Icon icon={RedisGraphDark} {...props} isSvg />
-)
