--- redisinsight/ui/src/components/base/icons/modules/RedisGraphLightIcon.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-import RedisGraphLight from 'uiSrc/assets/img/modules/RedisGraphLight.svg?react'
-import { Icon, IconProps } from 'uiSrc/components/base/icons'
-
-export const RedisGraphLightIcon = (props: IconProps) => (
-  <Icon icon={RedisGraphLight} {...props} isSvg />
-)
