--- redisinsight/ui/src/components/base/icons/modules/RedisJSONDarkIcon.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-import RedisJSONDark from 'uiSrc/assets/img/modules/RedisJSONDark.svg?react'
-import { Icon, IconProps } from 'uiSrc/components/base/icons'
-
-export const RedisJSONDarkIcon = (props: IconProps) => (
-  <Icon icon={RedisJSONDark} {...props} isSvg />
-)
