--- redisinsight/ui/src/components/base/icons/modules/RedisJSONLightIcon.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-import RedisJSONLight from 'uiSrc/assets/img/modules/RedisJSONLight.svg?react'
-import { Icon, IconProps } from 'uiSrc/components/base/icons'
-
-export const RedisJSONLightIcon = (props: IconProps) => (
-  <Icon icon={RedisJSONLight} {...props} isSvg />
-)
