--- redisinsight/ui/src/components/base/icons/modules/RedisSearchDarkIcon.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-import RedisSearchDark from 'uiSrc/assets/img/modules/RedisSearchDark.svg?react'
-import { Icon, IconProps } from 'uiSrc/components/base/icons'
-
-export const RedisSearchDarkIcon = (props: IconProps) => (
-  <Icon icon={RedisSearchDark} {...props} isSvg />
-)
