--- redisinsight/ui/src/components/base/icons/modules/RedisSearchLightIcon.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-import RedisSearchLight from 'uiSrc/assets/img/modules/RedisSearchLight.svg?react'
-import { Icon, IconProps } from 'uiSrc/components/base/icons'
-
-export const RedisSearchLightIcon = (props: IconProps) => (
-  <Icon icon={RedisSearchLight} {...props} isSvg />
-)
