--- redisinsight/ui/src/components/base/icons/modules/RedisTimeSeriesDarkIcon.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-import RedisTimeSeriesDark from 'uiSrc/assets/img/modules/RedisTimeSeriesDark.svg?react'
-import { Icon, IconProps } from 'uiSrc/components/base/icons'
-
-export const RedisTimeSeriesDarkIcon = (props: IconProps) => (
-  <Icon icon={RedisTimeSeriesDark} {...props} isSvg />
-)
