--- redisinsight/ui/src/components/base/icons/modules/RedisTimeSeriesLightIcon.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-import RedisTimeSeriesLight from 'uiSrc/assets/img/modules/RedisTimeSeriesLight.svg?react'
-import { Icon, IconProps } from 'uiSrc/components/base/icons'
-
-export const RedisTimeSeriesLightIcon = (props: IconProps) => (
-  <Icon icon={RedisTimeSeriesLight} {...props} isSvg />
-)
