--- redisinsight/ui/src/components/base/icons/modules/UnknownDarkIcon.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-import UnknownDark from 'uiSrc/assets/img/modules/UnknownDark.svg?react'
-import { Icon, IconProps } from 'uiSrc/components/base/icons'
-
-export const UnknownDarkIcon = (props: IconProps) => (
-  <Icon icon={UnknownDark} {...props} isSvg />
-)
