--- redisinsight/ui/src/components/base/icons/modules/UnknownLightIcon.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-import UnknownLight from 'uiSrc/assets/img/modules/UnknownLight.svg?react'
-import { Icon, IconProps } from 'uiSrc/components/base/icons'
-
-export const UnknownLightIcon = (props: IconProps) => (
-  <Icon icon={UnknownLight} {...props} isSvg />
-)
