--- redisinsight/ui/src/components/base/icons/modules/index.ts
@@ -1,22 +1,0 @@
-export { RedisAIDarkIcon } from './RedisAIDarkIcon'
-export { RedisAILightIcon } from './RedisAILightIcon'
-export { RedisBloomDarkIcon } from './RedisBloomDarkIcon'
-export { RedisBloomLightIcon } from './RedisBloomLightIcon'
-export { RedisGears2DarkIcon } from './RedisGears2DarkIcon'
-export { RedisGears2LightIcon } from './RedisGears2LightIcon'
-export { RedisGearsDarkIcon } from './RedisGearsDarkIcon'
-export { RedisGearsLightIcon } from './RedisGearsLightIcon'
-export { RedisGraphDarkIcon } from './RedisGraphDarkIcon'
-export { RedisGraphLightIcon } from './RedisGraphLightIcon'
-export { RedisJSONDarkIcon } from './RedisJSONDarkIcon'
-export { RedisJSONLightIcon } from './RedisJSONLightIcon'
-export { RedisSearchDarkIcon } from './RedisSearchDarkIcon'
-export { RedisSearchLightIcon } from './RedisSearchLightIcon'
-export { RedisTimeSeriesDarkIcon } from './RedisTimeSeriesDarkIcon'
-export { RedisTimeSeriesLightIcon } from './RedisTimeSeriesLightIcon'
-export { RediStackDarkMinIcon } from './RediStackDarkMinIcon'
-export { RediStackLightMinIcon } from './RediStackLightMinIcon'
-export { RediStackDarkLogoIcon } from './RediStackDarkLogoIcon'
-export { RediStackLightLogoIcon } from './RediStackLightLogoIcon'
-export { UnknownDarkIcon } from './UnknownDarkIcon'
-export { UnknownLightIcon } from './UnknownLightIcon'
