+++ redisinsight/ui/src/components/base/icons/modules/index.tsx
@@ -1,0 +1,113 @@
+import React from 'react'
+import { Icon, IconProps } from 'uiSrc/components/base/icons'
+
+import RediStackDarkLogoSvg from 'uiSrc/assets/img/modules/redistack/RedisStackLogoDark.svg?react'
+import RediStackDarkMinSvg from 'uiSrc/assets/img/modules/redistack/RediStackDark-min.svg?react'
+import RediStackLightLogoSvg from 'uiSrc/assets/img/modules/redistack/RedisStackLogoLight.svg?react'
+import RediStackLightMinLight from 'uiSrc/assets/img/modules/redistack/RediStackLight-min.svg?react'
+import RedisAIDark from 'uiSrc/assets/img/modules/RedisAIDark.svg?react'
+import RedisAILight from 'uiSrc/assets/img/modules/RedisAILight.svg?react'
+import RedisBloomDark from 'uiSrc/assets/img/modules/RedisBloomDark.svg?react'
+import RedisBloomLight from 'uiSrc/assets/img/modules/RedisBloomLight.svg?react'
+import RedisGears2Dark from 'uiSrc/assets/img/modules/RedisGears2Dark.svg?react'
+import RedisGears2Light from 'uiSrc/assets/img/modules/RedisGears2Light.svg?react'
+import RedisGearsDark from 'uiSrc/assets/img/modules/RedisGearsDark.svg?react'
+import RedisGearsLight from 'uiSrc/assets/img/modules/RedisGearsLight.svg?react'
+import RedisGraphDark from 'uiSrc/assets/img/modules/RedisGraphDark.svg?react'
+import RedisGraphLight from 'uiSrc/assets/img/modules/RedisGraphLight.svg?react'
+import RedisJSONDark from 'uiSrc/assets/img/modules/RedisJSONDark.svg?react'
+import RedisJSONLight from 'uiSrc/assets/img/modules/RedisJSONLight.svg?react'
+import RedisSearchDark from 'uiSrc/assets/img/modules/RedisSearchDark.svg?react'
+import RedisSearchLight from 'uiSrc/assets/img/modules/RedisSearchLight.svg?react'
+import RedisTimeSeriesDark from 'uiSrc/assets/img/modules/RedisTimeSeriesDark.svg?react'
+import RedisTimeSeriesLight from 'uiSrc/assets/img/modules/RedisTimeSeriesLight.svg?react'
+import UnknownDark from 'uiSrc/assets/img/modules/UnknownDark.svg?react'
+import UnknownLight from 'uiSrc/assets/img/modules/UnknownLight.svg?react'
+
+export const RedisAIDarkIcon = (props: IconProps) => (
+  <Icon icon={RedisAIDark} {...props} isSvg />
+)
+
+export const RedisAILightIcon = (props: IconProps) => (
+  <Icon icon={RedisAILight} {...props} isSvg />
+)
+
+export const RedisBloomDarkIcon = (props: IconProps) => (
+  <Icon icon={RedisBloomDark} {...props} isSvg />
+)
+
+export const RedisBloomLightIcon = (props: IconProps) => (
+  <Icon icon={RedisBloomLight} {...props} isSvg />
+)
+
+export const RedisGears2DarkIcon = (props: IconProps) => (
+  <Icon icon={RedisGears2Dark} {...props} isSvg />
+)
+
+export const RedisGears2LightIcon = (props: IconProps) => (
+  <Icon icon={RedisGears2Light} {...props} isSvg />
+)
+
+export const RedisGearsDarkIcon = (props: IconProps) => (
+  <Icon icon={RedisGearsDark} {...props} isSvg />
+)
+
+export const RedisGearsLightIcon = (props: IconProps) => (
+  <Icon icon={RedisGearsLight} {...props} isSvg />
+)
+
+export const RedisGraphDarkIcon = (props: IconProps) => (
+  <Icon icon={RedisGraphDark} {...props} isSvg />
+)
+
+export const RedisGraphLightIcon = (props: IconProps) => (
+  <Icon icon={RedisGraphLight} {...props} isSvg />
+)
+
+export const RedisJSONDarkIcon = (props: IconProps) => (
+  <Icon icon={RedisJSONDark} {...props} isSvg />
+)
+
+export const RedisJSONLightIcon = (props: IconProps) => (
+  <Icon icon={RedisJSONLight} {...props} isSvg />
+)
+
+export const RedisSearchDarkIcon = (props: IconProps) => (
+  <Icon icon={RedisSearchDark} {...props} isSvg />
+)
+
+export const RedisSearchLightIcon = (props: IconProps) => (
+  <Icon icon={RedisSearchLight} {...props} isSvg />
+)
+
+export const RediStackDarkLogoIcon = (props: IconProps) => (
+  <Icon icon={RediStackDarkLogoSvg} {...props} isSvg />
+)
+
+export const RediStackDarkMinIcon = (props: IconProps) => (
+  <Icon icon={RediStackDarkMinSvg} {...props} isSvg />
+)
+
+export const RediStackLightLogoIcon = (props: IconProps) => (
+  <Icon icon={RediStackLightLogoSvg} {...props} isSvg />
+)
+
+export const RediStackLightMinIcon = (props: IconProps) => (
+  <Icon icon={RediStackLightMinLight} {...props} isSvg />
+)
+
+export const RedisTimeSeriesDarkIcon = (props: IconProps) => (
+  <Icon icon={RedisTimeSeriesDark} {...props} isSvg />
+)
+
+export const RedisTimeSeriesLightIcon = (props: IconProps) => (
+  <Icon icon={RedisTimeSeriesLight} {...props} isSvg />
+)
+
+export const UnknownDarkIcon = (props: IconProps) => (
+  <Icon icon={UnknownDark} {...props} isSvg />
+)
+
+export const UnknownLightIcon = (props: IconProps) => (
+  <Icon icon={UnknownLight} {...props} isSvg />
+)
